requirements.txt
streamlit
pytube
openai-whisper
moviepy
opencv-python
numpy
pillow